refactor(agent): extract clawdstrike-fs-policy-paths crate
Move ~850 lines of pure OS persistence-target path validators out of
apps/agent/src-tauri/src/api_server/response_actions.rs into a new
clawdstrike-fs-policy-paths crate. The crate exposes a single public
predicate, path_is_bounded_persistence_target, that covers macOS
LaunchAgents/LaunchDaemons, systemd user/system units and drop-ins,
XDG autostart entries, KDE Plasma env + autostart scripts, shell
startup files (bash/zsh/fish), system profile.d, user cron spools,
system cron drop-ins, and Chromium/Firefox extension manifests.

The functions are pure (Path -> bool, zero IO, zero AgentApiState
references) and previously had a literal duplicate footprint planned
for the control-plane validator. Co-locating them removes the
agent-enforcer / control-plane drift risk identified in the
architecture synthesis.

Tests for the path validators (~373 lines, 8 functions) moved with
the code from api_server/tests/response_and_settings/part_3.rs into
the new crate. The aggregator stays reachable from the agent through
a pub(crate) use re-export in response_actions.rs, so existing call
sites (disable_persistence_target_path, validate_disable_persistence_*
and default_detection_candidate_action in api_server.rs) work
unchanged.

No behaviour changes. cargo build/test pass for both the new crate
and the agent binary; pre-existing agent test failures
(macos::collector::*, agent_edr_findings_redacts_*) are unaffected.
