+++ apps/agent/src-tauri/Cargo.toml
@@ -61,6 +61,7 @@ which = "6"
 hush-core = { path = "../../../crates/libs/hush-core" }
 clawdstrike-policy-event = { path = "../../../crates/libs/clawdstrike-policy-event", default-features = false, features = ["simulate"] }
 clawdstrike-hushd-config = { path = "../../../crates/libs/clawdstrike-hushd-config" }
+clawdstrike-fs-policy-paths = { path = "../../../crates/libs/clawdstrike-fs-policy-paths" }
 base64 = "0.22"
 
 # NATS / Spine (adaptive SDR)
