+++ apps/agent/src-tauri/src/api_server/response_actions.rs
@@ -1,5 +1,7 @@
 use super::*;
 
+pub(crate) use clawdstrike_fs_policy_paths::path_is_bounded_persistence_target;
+
 pub(crate) fn supported_edr_simulation_action(action: &EndpointDecisionAction) -> bool {
     matches!(
         action,
@@ -2791,856 +2793,6 @@ fn path_is_bounded_quarantine_source(path: &FsPath) -> bool {
         || normalized.contains("/node_modules/")
         || normalized.contains("/target/debug/")
         || normalized.contains("/target/release/")
-}
-
-pub(crate) fn path_is_bounded_persistence_target(path: &FsPath) -> bool {
-    path_is_bounded_launch_persistence_target(path)
-        || path_is_bounded_systemd_user_persistence_target(path)
-        || path_is_bounded_systemd_system_persistence_target(path)
-        || path_is_bounded_systemd_user_dropin_persistence_target(path)
-        || path_is_bounded_systemd_system_dropin_persistence_target(path)
-        || path_is_bounded_xdg_autostart_persistence_target(path)
-        || path_is_bounded_plasma_env_persistence_target(path)
-        || path_is_bounded_kde_autostart_script_persistence_target(path)
-        || path_is_bounded_shell_startup_persistence_target(path)
-        || path_is_bounded_profile_d_persistence_target(path)
-        || path_is_bounded_cron_persistence_target(path)
-        || path_is_bounded_system_cron_dropin_persistence_target(path)
-        || path_is_bounded_browser_extension_manifest_target(path)
-}
-
-fn path_is_bounded_launch_persistence_target(path: &FsPath) -> bool {
-    let normalized = path.display().to_string().replace('\\', "/");
-    if !normalized.ends_with(".plist")
-        || !normalized.contains("/Library/LaunchAgents/")
-            && !normalized.contains("/Library/LaunchDaemons/")
-    {
-        return false;
-    }
-    if let Some(file_name) = path.file_name().and_then(|value| value.to_str()) {
-        let normalized_file = file_name.to_ascii_lowercase();
-        if normalized_file.starts_with("com.apple.")
-            || normalized_file.starts_with("com.clawdstrike.")
-            || normalized_file.starts_with("io.clawdstrike.")
-        {
-            return false;
-        }
-    }
-    let temp_dir = std::env::temp_dir()
-        .display()
-        .to_string()
-        .replace('\\', "/");
-    normalized.starts_with(temp_dir.as_str())
-        || normalized.starts_with("/tmp/")
-        || normalized.starts_with("/private/tmp/")
-        || normalized.starts_with("/var/folders/")
-        || normalized.starts_with("/Library/LaunchAgents/")
-        || normalized.starts_with("/Library/LaunchDaemons/")
-        || (normalized.starts_with("/Users/") && !normalized.contains("/../"))
-}
-
-fn path_is_bounded_systemd_user_persistence_target(path: &FsPath) -> bool {
-    let normalized = path.display().to_string().replace('\\', "/");
-    if normalized.contains("/../") {
-        return false;
-    }
-    let Some(file_name) = path.file_name().and_then(|value| value.to_str()) else {
-        return false;
-    };
-    if !systemd_unit_file_name_is_safe(file_name) {
-        return false;
-    }
-
-    let temp_dir = std::env::temp_dir()
-        .display()
-        .to_string()
-        .replace('\\', "/");
-    let temp_like = normalized.starts_with(temp_dir.as_str())
-        || normalized.starts_with("/tmp/")
-        || normalized.starts_with("/private/tmp/")
-        || normalized.starts_with("/var/folders/");
-    let user_home_like = normalized.starts_with("/home/") || normalized.starts_with("/Users/");
-    if !temp_like && !user_home_like {
-        return false;
-    }
-
-    let components = normalized
-        .trim_start_matches('/')
-        .split('/')
-        .filter(|component| !component.is_empty())
-        .collect::<Vec<_>>();
-    systemd_user_unit_components_are_bounded(&components, file_name)
-}
-
-fn systemd_user_unit_components_are_bounded(components: &[&str], file_name: &str) -> bool {
-    if components.len() < 6 {
-        return false;
-    }
-    let base = components.len() - 6;
-    let home_component = components[base];
-    let user = components[base + 1];
-    matches!(home_component, "home" | "Users")
-        && cron_spool_user_name_is_safe(user)
-        && components[base + 2] == ".config"
-        && components[base + 3] == "systemd"
-        && components[base + 4] == "user"
-        && components[base + 5] == file_name
-}
-
-fn path_is_bounded_systemd_user_dropin_persistence_target(path: &FsPath) -> bool {
-    let normalized = path.display().to_string().replace('\\', "/");
-    if normalized.contains("/../") {
-        return false;
-    }
-    let Some(file_name) = path.file_name().and_then(|value| value.to_str()) else {
-        return false;
-    };
-    if !systemd_dropin_file_name_is_safe(file_name) {
-        return false;
-    }
-
-    let temp_dir = std::env::temp_dir()
-        .display()
-        .to_string()
-        .replace('\\', "/");
-    let temp_like = normalized.starts_with(temp_dir.as_str())
-        || normalized.starts_with("/tmp/")
-        || normalized.starts_with("/private/tmp/")
-        || normalized.starts_with("/var/folders/");
-    let user_home_like = normalized.starts_with("/home/") || normalized.starts_with("/Users/");
-    if !temp_like && !user_home_like {
-        return false;
-    }
-
-    let components = normalized
-        .trim_start_matches('/')
-        .split('/')
-        .filter(|component| !component.is_empty())
-        .collect::<Vec<_>>();
-    systemd_user_dropin_components_are_bounded(&components, file_name)
-}
-
-fn systemd_user_dropin_components_are_bounded(components: &[&str], file_name: &str) -> bool {
-    if components.len() < 7 {
-        return false;
-    }
-    let base = components.len() - 7;
-    let home_component = components[base];
-    let user = components[base + 1];
-    let unit_dir = components[base + 5];
-    matches!(home_component, "home" | "Users")
-        && cron_spool_user_name_is_safe(user)
-        && components[base + 2] == ".config"
-        && components[base + 3] == "systemd"
-        && components[base + 4] == "user"
-        && systemd_unit_dropin_dir_name_is_safe(unit_dir)
-        && components[base + 6] == file_name
-}
-
-fn path_is_bounded_systemd_system_persistence_target(path: &FsPath) -> bool {
-    let normalized = path.display().to_string().replace('\\', "/");
-    if normalized.contains("/../") {
-        return false;
-    }
-    let Some(file_name) = path.file_name().and_then(|value| value.to_str()) else {
-        return false;
-    };
-    if !systemd_system_unit_file_name_is_safe(file_name) {
-        return false;
-    }
-
-    let temp_dir = std::env::temp_dir()
-        .display()
-        .to_string()
-        .replace('\\', "/");
-    let temp_like = normalized.starts_with(temp_dir.as_str())
-        || normalized.starts_with("/tmp/")
-        || normalized.starts_with("/private/tmp/")
-        || normalized.starts_with("/var/folders/");
-    let system_unit_like = normalized.starts_with("/etc/systemd/system/");
-    if !temp_like && !system_unit_like {
-        return false;
-    }
-
-    let components = normalized
-        .trim_start_matches('/')
-        .split('/')
-        .filter(|component| !component.is_empty())
-        .collect::<Vec<_>>();
-    cron_spool_components_end_with(&components, &["etc", "systemd", "system", file_name])
-}
-
-fn path_is_bounded_systemd_system_dropin_persistence_target(path: &FsPath) -> bool {
-    let normalized = path.display().to_string().replace('\\', "/");
-    if normalized.contains("/../") {
-        return false;
-    }
-    let Some(file_name) = path.file_name().and_then(|value| value.to_str()) else {
-        return false;
-    };
-    if !systemd_dropin_file_name_is_safe(file_name) {
-        return false;
-    }
-
-    let temp_dir = std::env::temp_dir()
-        .display()
-        .to_string()
-        .replace('\\', "/");
-    let temp_like = normalized.starts_with(temp_dir.as_str())
-        || normalized.starts_with("/tmp/")
-        || normalized.starts_with("/private/tmp/")
-        || normalized.starts_with("/var/folders/");
-    let system_unit_like = normalized.starts_with("/etc/systemd/system/");
-    if !temp_like && !system_unit_like {
-        return false;
-    }
-
-    let components = normalized
-        .trim_start_matches('/')
-        .split('/')
-        .filter(|component| !component.is_empty())
-        .collect::<Vec<_>>();
-    systemd_system_dropin_components_are_bounded(&components, file_name)
-}
-
-fn systemd_system_dropin_components_are_bounded(components: &[&str], file_name: &str) -> bool {
-    if components.len() < 5 {
-        return false;
-    }
-    let base = components.len() - 5;
-    let unit_dir = components[base + 3];
-    components[base] == "etc"
-        && components[base + 1] == "systemd"
-        && components[base + 2] == "system"
-        && systemd_system_unit_dropin_dir_name_is_safe(unit_dir)
-        && components[base + 4] == file_name
-}
-
-fn systemd_unit_file_name_is_safe(file_name: &str) -> bool {
-    let Some((stem, extension)) = file_name.rsplit_once('.') else {
-        return false;
-    };
-    matches!(extension, "service" | "timer" | "socket")
-        && !stem.is_empty()
-        && stem.len() <= 128
-        && !stem.starts_with('.')
-        && !stem.eq_ignore_ascii_case("clawdstrike")
-        && !stem.to_ascii_lowercase().starts_with("clawdstrike.")
-        && stem
-            .bytes()
-            .all(|byte| byte.is_ascii_alphanumeric() || matches!(byte, b'-' | b'_' | b'.' | b'@'))
-}
-
-fn systemd_unit_dropin_dir_name_is_safe(dir_name: &str) -> bool {
-    dir_name
-        .strip_suffix(".d")
-        .is_some_and(systemd_unit_file_name_is_safe)
-}
-
-fn systemd_system_unit_dropin_dir_name_is_safe(dir_name: &str) -> bool {
-    dir_name
-        .strip_suffix(".d")
-        .is_some_and(systemd_system_unit_file_name_is_safe)
-}
-
-fn systemd_dropin_file_name_is_safe(file_name: &str) -> bool {
-    let Some((stem, extension)) = file_name.rsplit_once('.') else {
-        return false;
-    };
-    extension == "conf"
-        && !stem.is_empty()
-        && stem.len() <= 128
-        && !stem.starts_with('.')
-        && !stem.eq_ignore_ascii_case("clawdstrike")
-        && !stem.to_ascii_lowercase().starts_with("clawdstrike.")
-        && stem
-            .bytes()
-            .all(|byte| byte.is_ascii_alphanumeric() || matches!(byte, b'-' | b'_' | b'.' | b'@'))
-}
-
-fn systemd_system_unit_file_name_is_safe(file_name: &str) -> bool {
-    if !systemd_unit_file_name_is_safe(file_name) {
-        return false;
-    }
-    let stem = file_name
-        .rsplit_once('.')
-        .map(|(stem, _)| stem)
-        .unwrap_or(file_name);
-    let normalized_stem = stem.to_ascii_lowercase();
-    if normalized_stem.starts_with("systemd")
-        || normalized_stem.starts_with("getty")
-        || normalized_stem.starts_with("serial-getty")
-        || normalized_stem.starts_with("clawdstrike")
-        || normalized_stem.starts_with("com.clawdstrike")
-        || normalized_stem.starts_with("io.clawdstrike")
-    {
-        return false;
-    }
-    !matches!(
-        normalized_stem.as_str(),
-        "ssh"
-            | "sshd"
-            | "cron"
-            | "crond"
-            | "dbus"
-            | "networkmanager"
-            | "networking"
-            | "containerd"
-            | "docker"
-            | "kubelet"
-            | "login"
-            | "sudo"
-            | "polkit"
-    )
-}
-
-fn path_is_bounded_xdg_autostart_persistence_target(path: &FsPath) -> bool {
-    let normalized = path.display().to_string().replace('\\', "/");
-    if normalized.contains("/../") {
-        return false;
-    }
-    let Some(file_name) = path.file_name().and_then(|value| value.to_str()) else {
-        return false;
-    };
-    if !xdg_autostart_desktop_file_name_is_safe(file_name) {
-        return false;
-    }
-
-    let temp_dir = std::env::temp_dir()
-        .display()
-        .to_string()
-        .replace('\\', "/");
-    let temp_like = normalized.starts_with(temp_dir.as_str())
-        || normalized.starts_with("/tmp/")
-        || normalized.starts_with("/private/tmp/")
-        || normalized.starts_with("/var/folders/");
-    let user_home_like = normalized.starts_with("/home/") || normalized.starts_with("/Users/");
-    let system_autostart_like = normalized.starts_with("/etc/xdg/autostart/");
-    if !temp_like && !user_home_like && !system_autostart_like {
-        return false;
-    }
-
-    let components = normalized
-        .trim_start_matches('/')
-        .split('/')
-        .filter(|component| !component.is_empty())
-        .collect::<Vec<_>>();
-    xdg_user_autostart_components_are_bounded(&components, file_name)
-        || xdg_system_autostart_components_are_bounded(&components, file_name)
-}
-
-fn xdg_user_autostart_components_are_bounded(components: &[&str], file_name: &str) -> bool {
-    if components.len() < 5 {
-        return false;
-    }
-    let base = components.len() - 5;
-    let home_component = components[base];
-    let user = components[base + 1];
-    matches!(home_component, "home" | "Users")
-        && cron_spool_user_name_is_safe(user)
-        && components[base + 2] == ".config"
-        && components[base + 3] == "autostart"
-        && components[base + 4] == file_name
-}
-
-fn xdg_system_autostart_components_are_bounded(components: &[&str], file_name: &str) -> bool {
-    cron_spool_components_end_with(components, &["etc", "xdg", "autostart", file_name])
-        && xdg_system_autostart_desktop_file_name_is_safe(file_name)
-}
-
-fn xdg_autostart_desktop_file_name_is_safe(file_name: &str) -> bool {
-    let Some((stem, extension)) = file_name.rsplit_once('.') else {
-        return false;
-    };
-    extension == "desktop"
-        && !stem.is_empty()
-        && stem.len() <= 128
-        && !stem.starts_with('.')
-        && !stem.eq_ignore_ascii_case("clawdstrike")
-        && !stem.to_ascii_lowercase().starts_with("clawdstrike.")
-        && stem
-            .bytes()
-            .all(|byte| byte.is_ascii_alphanumeric() || matches!(byte, b'-' | b'_' | b'.'))
-}
-
-fn xdg_system_autostart_desktop_file_name_is_safe(file_name: &str) -> bool {
-    if !xdg_autostart_desktop_file_name_is_safe(file_name) {
-        return false;
-    }
-    let stem = file_name
-        .rsplit_once('.')
-        .map(|(stem, _)| stem)
-        .unwrap_or(file_name)
-        .to_ascii_lowercase();
-    if stem.starts_with("org.gnome.")
-        || stem.starts_with("org.kde.")
-        || stem.starts_with("org.freedesktop.")
-        || stem.starts_with("gnome-keyring")
-        || stem.starts_with("clawdstrike")
-    {
-        return false;
-    }
-    !matches!(
-        stem.as_str(),
-        "polkit" | "dbus" | "ssh-agent" | "at-spi-dbus-bus" | "xdg-user-dirs" | "nm-applet"
-    )
-}
-
-fn path_is_bounded_plasma_env_persistence_target(path: &FsPath) -> bool {
-    let normalized = path.display().to_string().replace('\\', "/");
-    if normalized.contains("/../") {
-        return false;
-    }
-    let Some(file_name) = path.file_name().and_then(|value| value.to_str()) else {
-        return false;
-    };
-    if !plasma_env_script_file_name_is_safe(file_name) {
-        return false;
-    }
-
-    let temp_dir = std::env::temp_dir()
-        .display()
-        .to_string()
-        .replace('\\', "/");
-    let temp_like = normalized.starts_with(temp_dir.as_str())
-        || normalized.starts_with("/tmp/")
-        || normalized.starts_with("/private/tmp/")
-        || normalized.starts_with("/var/folders/");
-    let user_home_like = normalized.starts_with("/home/") || normalized.starts_with("/Users/");
-    if !temp_like && !user_home_like {
-        return false;
-    }
-
-    let components = normalized
-        .trim_start_matches('/')
-        .split('/')
-        .filter(|component| !component.is_empty())
-        .collect::<Vec<_>>();
-    plasma_env_components_are_bounded(&components, file_name)
-}
-
-fn plasma_env_components_are_bounded(components: &[&str], file_name: &str) -> bool {
-    if components.len() < 6 {
-        return false;
-    }
-    let base = components.len() - 6;
-    let home_component = components[base];
-    let user = components[base + 1];
-    matches!(home_component, "home" | "Users")
-        && cron_spool_user_name_is_safe(user)
-        && components[base + 2] == ".config"
-        && components[base + 3] == "plasma-workspace"
-        && components[base + 4] == "env"
-        && components[base + 5] == file_name
-}
-
-fn plasma_env_script_file_name_is_safe(file_name: &str) -> bool {
-    let Some((stem, extension)) = file_name.rsplit_once('.') else {
-        return false;
-    };
-    extension == "sh"
-        && !stem.is_empty()
-        && stem.len() <= 128
-        && !stem.starts_with('.')
-        && !stem.eq_ignore_ascii_case("clawdstrike")
-        && !stem.to_ascii_lowercase().starts_with("clawdstrike.")
-        && stem
-            .bytes()
-            .all(|byte| byte.is_ascii_alphanumeric() || matches!(byte, b'-' | b'_' | b'.'))
-}
-
-fn path_is_bounded_kde_autostart_script_persistence_target(path: &FsPath) -> bool {
-    let normalized = path.display().to_string().replace('\\', "/");
-    if normalized.contains("/../") {
-        return false;
-    }
-    let Some(file_name) = path.file_name().and_then(|value| value.to_str()) else {
-        return false;
-    };
-    if !plasma_env_script_file_name_is_safe(file_name) {
-        return false;
-    }
-
-    let temp_dir = std::env::temp_dir()
-        .display()
-        .to_string()
-        .replace('\\', "/");
-    let temp_like = normalized.starts_with(temp_dir.as_str())
-        || normalized.starts_with("/tmp/")
-        || normalized.starts_with("/private/tmp/")
-        || normalized.starts_with("/var/folders/");
-    let user_home_like = normalized.starts_with("/home/") || normalized.starts_with("/Users/");
-    if !temp_like && !user_home_like {
-        return false;
-    }
-
-    let components = normalized
-        .trim_start_matches('/')
-        .split('/')
-        .filter(|component| !component.is_empty())
-        .collect::<Vec<_>>();
-    kde_autostart_script_components_are_bounded(&components, file_name)
-}
-
-fn kde_autostart_script_components_are_bounded(components: &[&str], file_name: &str) -> bool {
-    if components.len() < 5 {
-        return false;
-    }
-    let base = components.len() - 5;
-    let home_component = components[base];
-    let user = components[base + 1];
-    matches!(home_component, "home" | "Users")
-        && cron_spool_user_name_is_safe(user)
-        && components[base + 2] == ".config"
-        && components[base + 3] == "autostart-scripts"
-        && components[base + 4] == file_name
-}
-
-fn path_is_bounded_shell_startup_persistence_target(path: &FsPath) -> bool {
-    let normalized = path.display().to_string().replace('\\', "/");
-    if normalized.contains("/../") {
-        return false;
-    }
-    let temp_dir = std::env::temp_dir()
-        .display()
-        .to_string()
-        .replace('\\', "/");
-    let temp_like = normalized.starts_with(temp_dir.as_str())
-        || normalized.starts_with("/tmp/")
-        || normalized.starts_with("/private/tmp/")
-        || normalized.starts_with("/var/folders/");
-    let user_home_like = normalized.starts_with("/Users/") || normalized.starts_with("/home/");
-    if !temp_like && !user_home_like {
-        return false;
-    }
-
-    let components = normalized
-        .trim_start_matches('/')
-        .split('/')
-        .filter(|component| !component.is_empty())
-        .collect::<Vec<_>>();
-    shell_startup_components_are_bounded(&components)
-}
-
-fn shell_startup_components_are_bounded(components: &[&str]) -> bool {
-    let home_shell_file = if components.len() >= 3 {
-        let base = components.len() - 3;
-        let home_component = components[base];
-        let user = components[base + 1];
-        let file = components[base + 2];
-        matches!(home_component, "Users" | "home")
-            && cron_spool_user_name_is_safe(user)
-            && matches!(
-                file,
-                ".bashrc" | ".bash_profile" | ".zshrc" | ".zprofile" | ".profile"
-            )
-    } else {
-        false
-    };
-    let fish_config = if components.len() >= 5 {
-        let base = components.len() - 5;
-        let home_component = components[base];
-        let user = components[base + 1];
-        matches!(home_component, "Users" | "home")
-            && cron_spool_user_name_is_safe(user)
-            && components[base + 2] == ".config"
-            && components[base + 3] == "fish"
-            && components[base + 4] == "config.fish"
-    } else {
-        false
-    };
-    let fish_conf_d = if components.len() >= 6 {
-        let base = components.len() - 6;
-        let home_component = components[base];
-        let user = components[base + 1];
-        let file = components[base + 5];
-        matches!(home_component, "Users" | "home")
-            && cron_spool_user_name_is_safe(user)
-            && components[base + 2] == ".config"
-            && components[base + 3] == "fish"
-            && components[base + 4] == "conf.d"
-            && fish_conf_d_file_name_is_safe(file)
-    } else {
-        false
-    };
-
-    home_shell_file || fish_config || fish_conf_d
-}
-
-fn fish_conf_d_file_name_is_safe(file_name: &str) -> bool {
-    let Some((stem, extension)) = file_name.rsplit_once('.') else {
-        return false;
-    };
-    extension == "fish"
-        && !stem.is_empty()
-        && stem.len() <= 128
-        && !stem.starts_with('.')
-        && !stem.eq_ignore_ascii_case("clawdstrike")
-        && !matches!(stem.to_ascii_lowercase().as_str(), "config" | "conf")
-        && !stem.to_ascii_lowercase().starts_with("clawdstrike.")
-        && stem
-            .bytes()
-            .all(|byte| byte.is_ascii_alphanumeric() || matches!(byte, b'-' | b'_' | b'.'))
-}
-
-fn path_is_bounded_profile_d_persistence_target(path: &FsPath) -> bool {
-    let normalized = path.display().to_string().replace('\\', "/");
-    if normalized.contains("/../") {
-        return false;
-    }
-    let Some(file_name) = path.file_name().and_then(|value| value.to_str()) else {
-        return false;
-    };
-    if !profile_d_script_file_name_is_safe(file_name) {
-        return false;
-    }
-
-    let temp_dir = std::env::temp_dir()
-        .display()
-        .to_string()
-        .replace('\\', "/");
-    let temp_like = normalized.starts_with(temp_dir.as_str())
-        || normalized.starts_with("/tmp/")
-        || normalized.starts_with("/private/tmp/")
-        || normalized.starts_with("/var/folders/");
-    let system_profile_like = normalized.starts_with("/etc/profile.d/");
-    if !temp_like && !system_profile_like {
-        return false;
-    }
-
-    let components = normalized
-        .trim_start_matches('/')
-        .split('/')
-        .filter(|component| !component.is_empty())
-        .collect::<Vec<_>>();
-    cron_spool_components_end_with(&components, &["etc", "profile.d", file_name])
-}
-
-fn path_is_bounded_cron_persistence_target(path: &FsPath) -> bool {
-    let normalized = path.display().to_string().replace('\\', "/");
-    if normalized.contains("/../") {
-        return false;
-    }
-    let Some(user) = path.file_name().and_then(|value| value.to_str()) else {
-        return false;
-    };
-    if !cron_spool_user_name_is_safe(user) {
-        return false;
-    }
-
-    let temp_dir = std::env::temp_dir()
-        .display()
-        .to_string()
-        .replace('\\', "/");
-    let temp_like = normalized.starts_with(temp_dir.as_str())
-        || normalized.starts_with("/tmp/")
-        || normalized.starts_with("/private/tmp/")
-        || normalized.starts_with("/var/folders/");
-    let system_cron_like =
-        normalized.starts_with("/var/spool/cron/") || normalized.starts_with("/usr/lib/cron/tabs/");
-    if !temp_like && !system_cron_like {
-        return false;
-    }
-
-    let components = normalized
-        .trim_start_matches('/')
-        .split('/')
-        .filter(|component| !component.is_empty())
-        .collect::<Vec<_>>();
-    cron_spool_components_end_with(&components, &["var", "spool", "cron", "crontabs", user])
-        || cron_spool_components_end_with(&components, &["var", "spool", "cron", user])
-        || cron_spool_components_end_with(&components, &["usr", "lib", "cron", "tabs", user])
-}
-
-fn path_is_bounded_system_cron_dropin_persistence_target(path: &FsPath) -> bool {
-    let normalized = path.display().to_string().replace('\\', "/");
-    if normalized.contains("/../") {
-        return false;
-    }
-    let Some(file_name) = path.file_name().and_then(|value| value.to_str()) else {
-        return false;
-    };
-    if !cron_dropin_file_name_is_safe(file_name) {
-        return false;
-    }
-
-    let temp_dir = std::env::temp_dir()
-        .display()
-        .to_string()
-        .replace('\\', "/");
-    let temp_like = normalized.starts_with(temp_dir.as_str())
-        || normalized.starts_with("/tmp/")
-        || normalized.starts_with("/private/tmp/")
-        || normalized.starts_with("/var/folders/");
-    let system_cron_dropin_like = normalized.starts_with("/etc/cron.d/");
-    if !temp_like && !system_cron_dropin_like {
-        return false;
-    }
-
-    let components = normalized
-        .trim_start_matches('/')
-        .split('/')
-        .filter(|component| !component.is_empty())
-        .collect::<Vec<_>>();
-    cron_spool_components_end_with(&components, &["etc", "cron.d", file_name])
-}
-
-fn path_is_bounded_browser_extension_manifest_target(path: &FsPath) -> bool {
-    let normalized = path.display().to_string().replace('\\', "/");
-    if normalized.contains("/../") {
-        return false;
-    }
-    let Some(file_name) = path.file_name().and_then(|value| value.to_str()) else {
-        return false;
-    };
-    if !file_name.eq_ignore_ascii_case("manifest.json") {
-        return false;
-    }
-
-    let temp_dir = std::env::temp_dir()
-        .display()
-        .to_string()
-        .replace('\\', "/");
-    let allowed_root = normalized.starts_with(temp_dir.as_str())
-        || normalized.starts_with("/tmp/")
-        || normalized.starts_with("/private/tmp/")
-        || normalized.starts_with("/var/folders/")
-        || normalized.starts_with("/home/")
-        || normalized.starts_with("/Users/");
-    if !allowed_root {
-        return false;
-    }
-
-    let chromium_manifest = [
-        "/Library/Application Support/Google/Chrome/",
-        "/Library/Application Support/Chromium/",
-        "/Library/Application Support/Microsoft Edge/",
-        "/Library/Application Support/BraveSoftware/Brave-Browser/",
-    ]
-    .iter()
-    .any(|marker| normalized.contains(marker))
-        && normalized.contains("/Extensions/")
-        && chromium_browser_extension_manifest_components_are_bounded(&normalized);
-    let firefox_manifest = [
-        "/Library/Application Support/Firefox/Profiles/",
-        "/.mozilla/firefox/",
-    ]
-    .iter()
-    .any(|marker| firefox_browser_extension_manifest_components_are_bounded(&normalized, marker));
-
-    chromium_manifest || firefox_manifest
-}
-
-fn chromium_browser_extension_manifest_components_are_bounded(normalized: &str) -> bool {
-    let Some((_, extension_suffix)) = normalized.rsplit_once("/Extensions/") else {
-        return false;
-    };
-    let components = extension_suffix
-        .split('/')
-        .filter(|component| !component.is_empty())
-        .collect::<Vec<_>>();
-    components.len() == 3
-        && components.last() == Some(&"manifest.json")
-        && browser_extension_path_component_is_safe(components[0])
-        && browser_extension_path_component_is_safe(components[1])
-}
-
-fn firefox_browser_extension_manifest_components_are_bounded(
-    normalized: &str,
-    profile_marker: &str,
-) -> bool {
-    let Some((_, profile_suffix)) = normalized.rsplit_once(profile_marker) else {
-        return false;
-    };
-    let components = profile_suffix
-        .split('/')
-        .filter(|component| !component.is_empty())
-        .collect::<Vec<_>>();
-    components.len() == 4
-        && components[1] == "extensions"
-        && components[3].eq_ignore_ascii_case("manifest.json")
-        && browser_extension_path_component_is_safe(components[0])
-        && firefox_extension_path_component_is_safe(components[2])
-}
-
-fn browser_extension_path_component_is_safe(component: &str) -> bool {
-    !component.is_empty()
-        && component.len() <= 128
-        && !component.starts_with('.')
-        && component
-            .bytes()
-            .all(|byte| byte.is_ascii_alphanumeric() || matches!(byte, b'-' | b'_' | b'.'))
-}
-
-fn firefox_extension_path_component_is_safe(component: &str) -> bool {
-    !component.is_empty()
-        && component.len() <= 128
-        && !component.starts_with('.')
-        && component.bytes().all(|byte| {
-            byte.is_ascii_alphanumeric() || matches!(byte, b'-' | b'_' | b'.' | b'@' | b'{' | b'}')
-        })
-}
-
-fn cron_spool_components_end_with(components: &[&str], suffix: &[&str]) -> bool {
-    components.len() >= suffix.len()
-        && components[components.len() - suffix.len()..]
-            .iter()
-            .zip(suffix.iter())
-            .all(|(left, right)| left == right)
-}
-
-fn cron_spool_user_name_is_safe(user: &str) -> bool {
-    !user.is_empty()
-        && user.len() <= 64
-        && !user.starts_with('.')
-        && !matches!(user, "root" | "daemon" | "nobody")
-        && user
-            .bytes()
-            .all(|byte| byte.is_ascii_alphanumeric() || matches!(byte, b'-' | b'_' | b'.'))
-}
-
-fn cron_dropin_file_name_is_safe(file_name: &str) -> bool {
-    let normalized = file_name.to_ascii_lowercase();
-    !file_name.is_empty()
-        && file_name.len() <= 128
-        && !file_name.starts_with('.')
-        && !matches!(
-            normalized.as_str(),
-            "0hourly" | "anacron" | "cron" | "cronie" | "clawdstrike"
-        )
-        && !normalized.starts_with("clawdstrike.")
-        && !normalized.starts_with("com.clawdstrike.")
-        && file_name
-            .bytes()
-            .all(|byte| byte.is_ascii_alphanumeric() || matches!(byte, b'-' | b'_' | b'.'))
-}
-
-fn profile_d_script_file_name_is_safe(file_name: &str) -> bool {
-    let Some((stem, extension)) = file_name.rsplit_once('.') else {
-        return false;
-    };
-    let normalized_stem = stem.to_ascii_lowercase();
-    extension == "sh"
-        && !stem.is_empty()
-        && stem.len() <= 128
-        && !stem.starts_with('.')
-        && !stem.eq_ignore_ascii_case("clawdstrike")
-        && !normalized_stem.starts_with("clawdstrike.")
-        && !normalized_stem.starts_with("com.clawdstrike.")
-        && !matches!(
-            normalized_stem.as_str(),
-            "bash_completion"
-                | "colorgrep"
-                | "colorls"
-                | "lang"
-                | "locale"
-                | "proxy"
-                | "ssh-agent"
-                | "systemd"
-        )
-        && stem
-            .bytes()
-            .all(|byte| byte.is_ascii_alphanumeric() || matches!(byte, b'-' | b'_' | b'.'))
 }
 
 // quarantine_file_effect, disable_persistence_effect, suspend_process_tree_effect,
